stormwater_app/requirements.txt
streamlit>=1.32.0
python-docx>=1.1.0
Pillow>=10.2.0
pdfplumber>=0.11.0
python-dateutil>=2.9.0
anthropic>=0.40.0
python-dotenv>=1.0.0
plotly>=5.0.0
openpyxl>=3.1.0
httpx>=0.27.0
pandas>=2.0.0
